refactor(core): split transparent pipeline modules
